CLN: Consolidated and better-documented locking code
